getting unlostgit add -A! restructured folders
