+++ cart/cart.js
@@ -1,0 +1,6 @@
+//for dom manipulation on cart page
+export const cart = [
+    { id: '1', qty: 6 },
+    { id: '4', qty: 2 },
+];
+
