+++ cart/index.html
@@ -8,6 +8,7 @@
 </head>
 <body>
     <h1>Welcome to Cart</h1>
+    <a href="../">Back to Products</a>
     <table>
         <thead>
             <th>Product</th>
@@ -16,10 +17,12 @@
             <th>Total</th>
         </thead>
         <tbody id="table-body">
-            <tr>Chrysoprase</tr>
-            <tr>$10.00</tr>
-            <tr>2</tr>
-            <tr>$20.00</tr>
+            <!-- <tr>
+                <td>Chrysoprase</td>
+                <td>$10.00</td>
+                <td>2</td>
+                <td>$20.00</td>
+            </tr> -->
         </tbody>
         <tfoot>
             <td></td>
@@ -28,5 +31,7 @@
             <td>To Be Calculated</td>
         </tfoot>
     </table>
+    <button id="place-order">Place Order</button>
+    <script type="module" src="./cart.js"></script>
 </body>
 </html>
