--- data/cart-data.js
@@ -1,4 +1,0 @@
-export const cartContents = [
-    { id: '1', qty: 3},
-    { id: '4', qty: 2}
-];
