+++ data/cart-example-data.js
@@ -1,0 +1,4 @@
+export const cartContents = [
+    { id: '1', qty: 3 },
+    { id: '4', qty: 2 },
+];
