+++ index.html
@@ -13,7 +13,9 @@
 </head>
 
 <body>
-    <div id="rock-list">
+     <h1>Welcome to the Rock Shop!</h1>
+     <a href="./cart">Go to cart</a>
+    <div id="product-list">
         
     </div>
