+++ render-cart.js
@@ -1,0 +1,5 @@
+import { rocks } from './data/rocks.js';
+import { cart } from './data/cart-example-data.js';
+import { findById } from './utils.js';
+
+//create elements here
