+++ test/example.test.js
@@ -1,5 +1,6 @@
 import { renderRock } from '../render-rocks.js';
-import { rocks } from '../rock.js';
+import { rocks } from '../data/rocks.js';
+import { findById } from '../utils.js';
 // IMPORT MODULES under test here:
 // import { example } from '../example.js';
 
@@ -24,4 +25,15 @@ test('renderRock function', (expect) => {
     const chrysoprase = rocks[0];
     const actual = renderRock(chrysoprase).outerHTML;
     expect.equal(actual, expected);
+});
+
+test('findById should return the item matching the ID', (expect) => {
+    const expected = {
+        id: '1',
+        name: 'chrysoprase',
+        img: 'assets/chrysoprase.jpeg',
+        price: '5'
+    }
+    const actual = findById('1', rocks);
+    expect.deepEqual(actual, expected);
 });
